Merge latest FramePack changes with F1 support, and make it compatible with MPS (#4)
* Update README.md

* Update README.md

* Update README.md

* Update README.md

* prepare for sync

* Support FramePack-F1

FramePack-F1 is the framepack with forward-only sampling.

A GitHub discussion will be posted soon to describe it.

The model is trained with a new regulation approach for anti-drifting. This regulation will be uploaded to arxiv soon.

* Update README.md

* Update README.md

* Refactor data type handling in demo_gradio.py

Removed the --fp32 argument and simplified the data type assignment for transformer and encoder models to always use bfloat16 and float16, respectively. This change streamlines the code and ensures consistent model performance.

* Update F1 demo to work with MPS and add similar additional parameters as in the standard demo

---------

Co-authored-by: layerdiffusion <19834515+lllyasviel@users.noreply.github.com>